add orthograohic Camera & camera controller

+++ Hazel/src/Hazel/Events/ApplicationEvent.h
@@ -18,8 +18,8 @@ namespace Hazel {
 
 		}
 
-		inline float getWidth() { return m_Width; }
-		inline float getHeight() { return m_Height; }
+		inline float getWidth() { return (float)m_Width; }
+		inline float getHeight() { return (float)m_Height; }
 
 
 		std::string toString() const override
